+++ src/library/exceptions/library.exceptions.ts
@@ -1,4 +1,4 @@
-type ErrorName = "PERRMISION_DENIED" | "INVALID_BOOK";
+type ErrorName = "PERMISION_DENIED" | "INVALID_BOOK" | "BOOKING_ERROR";
 export class LibraryError extends Error {
   name: string;
   message: string;
