+++ src/library/service/library.service/__test__/library.service.test.ts
@@ -1,7 +1,6 @@
-import { Book } from "../../book/book";
-import { LibraryError } from "../../exceptions/library.exceptions";
-import { User } from "../../user/user";
-import { Booking } from "../components/booking";
+import { Book } from "../../../book/book";
+import { LibraryError } from "../../../exceptions/library.exceptions";
+import { User } from "../../../user/user";
 import { Library } from "../library.service";
 jest.mock("crypto", () => ({
   randomUUID: jest.fn().mockReturnValue("test-uuid"),
